feat: add useJounals
