+++ frontend/src/pages/Journals/Journals.tsx
@@ -1,11 +1,8 @@
 import CustomCalendar from '@/components/journals/CustomCalendar';
 import JournalCardList from '@/components/journals/JournalCardList';
-import { queryStringKeys } from '@/constants';
 import { NAV_HEIGHT, TOP_BAR_HEIGHT } from '@/constants/style';
 import { Dog } from '@/models/dog.model';
 import { Journal } from '@/models/journals';
-import { useEffect } from 'react';
-import { useLocation } from 'react-router-dom';
 
 const journals: Journal[] = [
     {
@@ -25,16 +22,8 @@ const dog: Dog = {
 };
 
 export default function Journals() {
-    const location = useLocation();
+    // const { journals } = useJournals();
 
-    useEffect(() => {
-        const params = new URLSearchParams(location.search);
-        const date = params.get(queryStringKeys.DATE);
-        if (date) {
-            // console.log(new Date(date));
-            // console.log(date);
-        }
-    }, [location.search]);
     return (
         <>
             <header></header>
